requirements.txt
requests~=2.28.2
telebot~=0.0.5
python-dotenv~=0.21.1
loguru~=0.6.0
backoff~=2.2.1
